Pull column logic out of config.go
